Moved auto curation script to title_tfidf.py
